Run patches in RblnPlatform.pre_register_and_update
Context:
We patch some vLLM modules for the cases where we need temporary fix for
things for which there is no proper customization support.

Problem:
We currently run the patches when vllm_rbln.worker is imported. But this
timing is not quite predictable, and can be too late for patching some
modules. While the current patches work well, I plan to add quantized
kernels (vLLM currently doesn't support OOT for this yet), which
requires patching some module before creating QuantizationConfig (due to
import order).

Solution:
Run patches in RblnPlatform.pre_register_and_update, which is
specifically desired for this purpose.
Also note that this is what vllm-ascend does:
<https://github.com/vllm-project/vllm-ascend/blob/875a86cbe9f8f88acb4fbcd84a4fbe89ac9bfd9d/docs/source/developer_guide/feature_guide/patch.md>